Implement modules for index.js
