+++ src/modules/results.js
@@ -1,0 +1,38 @@
+import {dateBuilder} from './weather';
+const api = {
+  key: '4cc8a926413ba21a3774022235df5269',
+  base: 'https://api.openweathermap.org/data/2.5/',
+};
+
+const searchbox = document.querySelector('.search-box');
+searchbox.addEventListener('keypress', setQuery);
+
+export function setQuery(evt) {
+  if (evt.keyCode === 13) {
+    getResults(searchbox.value);
+  }
+}
+
+async function getResults(query) {
+  await fetch(`${api.base}weather?q=${query}&units=metric&APPID=${api.key}`)
+    .then(weather => weather.json()).then(displayResults);
+}
+
+function displayResults(weather) {
+  const city = document.querySelector('.location .city');
+  city.innerText = `${weather.name}, ${weather.sys.country}`;
+
+  const now = new Date();
+  const date = document.querySelector('.location .date');
+  date.innerText = dateBuilder(now);
+
+  const temp = document.querySelector('.current .temp');
+  temp.innerHTML = `${Math.round(weather.main.temp)}<span>°C</span>`;
+
+  const weatherEl = document.querySelector('.current .weather');
+  weatherEl.innerText = weather.weather[0].main;
+
+  const hilow = document.querySelector('.hi-low');
+  hilow.innerText = `${Math.round(weather.main.temp_min)}°C / ${Math.round(weather.main.temp_max)}°C`;
+}
+
